requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
cffi==1.12.3
Click==7.0
dataclasses==0.6; python_version < '3.7'
greenlet==0.4.13
hypothesis==4.36.2
immutables==0.9
lark-parser==0.7.5
more-itertools==7.2.0
msgpack==0.6.1
pluggy==0.13.0
py==1.8.0
pytest==5.1.3
readline==6.2.4.1
six==1.12.0
typing-extensions==3.7.4
wcwidth==0.1.7